separate client/server ping_managers
