+++ client/src/connection/mod.rs
@@ -2,3 +2,5 @@
 pub mod connection;
 pub mod handshake_manager;
 pub mod io;
+pub mod ping_config;
+pub mod ping_manager;
