+++ client/src/connection/ping_config.rs
