+++ server/src/connection/mod.rs
@@ -1,5 +1,6 @@
 pub mod bandwidth_monitor;
-#[allow(clippy::module_inception)]
 pub mod connection;
 pub mod handshake_manager;
 pub mod io;
+pub mod ping_config;
+pub mod ping_manager;
