+++ server/src/server_config.rs
@@ -2,6 +2,8 @@ use std::default::Default;
 
 use naia_shared::ConnectionConfig;
 
+use crate::connection::ping_config::PingConfig;
+
 /// Contains Config properties which will be used by the Server
 #[derive(Clone)]
 pub struct ServerConfig {
@@ -10,6 +12,8 @@ pub struct ServerConfig {
     /// Determines whether to require that the Client send some auth message
     /// in order to connect.
     pub require_auth: bool,
+    /// Configuration used to monitor the ping & jitter on the network
+    pub ping: PingConfig,
 }
 
 impl Default for ServerConfig {
@@ -17,6 +21,7 @@ impl Default for ServerConfig {
         Self {
             connection: ConnectionConfig::default(),
             require_auth: true,
+            ping: PingConfig::default(),
         }
     }
 }
