+++ shared/src/connection/connection_config.rs
@@ -1,4 +1,3 @@
-use crate::PingConfig;
 use std::{default::Default, time::Duration};
 
 /// Contains Config properties which will be used by a Server or Client
@@ -13,8 +12,6 @@ pub struct ConnectionConfig {
     /// The duration over which to measure bandwidth. Set to None to avoid
     /// measure bandwidth at all.
     pub bandwidth_measure_duration: Option<Duration>,
-    /// Configuration used to monitor the ping & jitter on the network
-    pub ping: PingConfig,
 }
 
 impl ConnectionConfig {
@@ -23,13 +20,11 @@ impl ConnectionConfig {
         disconnection_timeout_duration: Duration,
         heartbeat_interval: Duration,
         bandwidth_measure_duration: Option<Duration>,
-        ping: PingConfig,
     ) -> Self {
         ConnectionConfig {
             disconnection_timeout_duration,
             heartbeat_interval,
             bandwidth_measure_duration,
-            ping,
         }
     }
 }
@@ -40,7 +35,6 @@ impl Default for ConnectionConfig {
             disconnection_timeout_duration: Duration::from_secs(30),
             heartbeat_interval: Duration::from_secs(4),
             bandwidth_measure_duration: None,
-            ping: PingConfig::default(),
         }
     }
 }
